pen_9

+++ src/components/AdminPage/ProductTable/index.jsx
@@ -1,25 +1,40 @@
-import React, { useEffect, useMemo, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { Table, Space, Button, Col } from 'antd';
-import { EditOutlined, DeleteOutlined } from '@ant-design/icons';
-import { actDeleteProduct, actFetchAllProduct, actUpdateProduct } from '../../../redux/features/products/productsSlice';
-import SearchManagement from '../SearchManagement';
-import { actSearchProduct } from '../../../redux/features/products/productsSlice';
-import { searchProduct } from '../../../apis/productApi';
+import { Table, Space, Button, Col, Modal } from 'antd';
+import { EditOutlined, DeleteOutlined, ExclamationCircleOutlined } from '@ant-design/icons';
+import { actDeleteProduct, actFetchAllProduct, actSearchProduct, actUpdateProduct } from '../../../redux/features/products/productsSlice';
+import SearchManagement from '../SearchManagement'
 const ProductTable = ({ onUpdateForm }) => {
+    const [query, setQuery] = useState('');
     const dispatch = useDispatch();
-    const { allProducts, isLoading } = useSelector(state => state.products);
+    const { allProducts, productSearch, isLoading } = useSelector(state => state.products);
     useEffect(() => {
-        dispatch(actFetchAllProduct());
-    }, []);
-    const reversedProducts = useMemo(() => {
-        return allProducts.slice().reverse();
-        //slice() để tạo bản sao của bản gốc để tránh làm thay đổi dữ liệu
-    }, [allProducts]);
-    const handleDeleteProduct = (id) => {
-        dispatch(actDeleteProduct(id));
-    };
+        dispatch(actSearchProduct(query));
+    }, [dispatch, query]);
+    useEffect(() => {
+        dispatch(actFetchAllProduct())
+    }, [])
+    const [isModalVisible, setIsModalVisible] = useState(false);
 
+    const handleDeleteProduct = (id) => {
+        Modal.confirm({
+            title: 'Bạn có chắc chắn muốn xóa sản phẩm này?',
+            icon: <ExclamationCircleOutlined />,
+            okText: 'Xóa',
+            cancelText: 'Hủy bỏ',
+            onOk: () => {
+              dispatch(actDeleteProduct(id));
+            },
+          });
+    };
+    const productFiltered = () => {
+        if (query) {
+            return productSearch.slice().reverse();
+        }
+        else {
+            return allProducts.slice().reverse();
+        }
+    }
     const handleEditProduct = (id) => {
         const product = allProducts.find(item => item.id === id);
         onUpdateForm(product);
@@ -33,18 +48,11 @@ const ProductTable = ({ onUpdateForm }) => {
     const handleTableChange = (pagination) => {
         setPagination(pagination);
     };
-    //HAM SEARCH
-    const [query, setQuery] = useState('');
-
-    const handleSearch = async(value) => {
-       setQuery(value);
-       const results = await searchProduct(value);
-       dispatch(actSearchProduct(results));  
-    };
-    const filteredProducts = useMemo(() => {
-        return reversedProducts;
-    }, [reversedProducts, query]);
-
+    //ham search
+    const handleSearch = (value) => {
+        setQuery(value);
+        console.log(value)
+    }
     const columns = [
         {
             title: 'Mã hàng',
@@ -97,13 +105,13 @@ const ProductTable = ({ onUpdateForm }) => {
     return (
         <div>
             <div style={{ display: 'flex', justifyContent: 'flex-end' }}>
-                <Col span={5}><SearchManagement value={query} placeholder='Search name or code' onSearch={handleSearch} onChange={(e)=>setQuery(e.target.value)}/></Col>
+                <Col span={5}><SearchManagement placeholder='search product...' onSearch={handleSearch} /></Col>
             </div>
             {isLoading ? (
                 <div>Loading...</div>
             ) : (
-                <Table columns={columns} dataSource={filteredProducts} rowKey="id" pagination={pagination}
-                    onChange={handleTableChange}  />
+                <Table columns={columns} dataSource={productFiltered()} rowKey="id" pagination={pagination}
+                    onChange={handleTableChange} />
             )}
         </div>
     );
